login admin as user as well as social media login features are moved to the frontend-api
- they use LoginAsUserFacade that needs FE API features to work properly (TokenAuthenticator and TokenFacade)
