+++ packages/framework/src/Model/Customer/User/CustomerUserUpdateDataFactory.php
@@ -173,38 +173,4 @@ class CustomerUserUpdateDataFactory implements CustomerUserUpdateDataFactoryInte
 
         return $deliveryAddressData;
     }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData $registrationData
-     * @return \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateData
-     */
-    public function createFromRegistrationData(RegistrationData $registrationData): CustomerUserUpdateData
-    {
-        $billingAddressData = $this->billingAddressDataFactory->create();
-        $billingAddressData->city = $registrationData->city;
-        $billingAddressData->street = $registrationData->street;
-        $billingAddressData->postcode = $registrationData->postcode;
-        $billingAddressData->country = $registrationData->country;
-        $billingAddressData->companyCustomer = $registrationData->companyCustomer;
-        $billingAddressData->companyName = $registrationData->companyName;
-        $billingAddressData->companyNumber = $registrationData->companyNumber;
-        $billingAddressData->companyTaxNumber = $registrationData->companyTaxNumber;
-        $billingAddressData->activated = $registrationData->activated;
-
-        $customerUserData = $this->customerUserDataFactory->createForDomainId($registrationData->domainId);
-        $customerUserData->createdAt = $registrationData->createdAt;
-        $customerUserData->email = $registrationData->email;
-        $customerUserData->lastName = $registrationData->lastName;
-        $customerUserData->password = $registrationData->password;
-        $customerUserData->firstName = $registrationData->firstName;
-        $customerUserData->telephone = $registrationData->telephone;
-        $customerUserData->newsletterSubscription = $registrationData->newsletterSubscription;
-
-        $customerUserUpdateData = $this->create();
-        $customerUserUpdateData->billingAddressData = $billingAddressData;
-        $customerUserUpdateData->customerUserData = $customerUserData;
-        $customerUserUpdateData->sendRegistrationMail = $registrationData->activated;
-
-        return $customerUserUpdateData;
-    }
 }
