+++ packages/framework/src/Model/Security/AdminLogoutHandler.php
@@ -28,7 +28,6 @@ class AdminLogoutHandler
     {
         $this->administratorLoginFacade->invalidateCurrentAdministratorLoginToken();
         $url = $this->router->generate('admin_login');
-        $request->getSession()->remove(LoginAsUserFacade::SESSION_LOGIN_AS);
         $request->getSession()->migrate();
 
         return new RedirectResponse($url);
