+++ packages/framework/src/Resources/views/Admin/Content/Customer/edit.html.twig
@@ -3,10 +3,4 @@
 {% block title %}- {{ 'Editing customer'|trans }} - {{ customerUser.fullName }}{% endblock %}
 {% block h1 %}
     {{ 'Editing customer'|trans }} - {{ customerUser.fullName }}
-    <a href="{{ ssoLoginAsUserUrl }}" target="_blank" class="btn-link-style btn-link-style--iconic-l">
-        <i class="svg svg-forward-page"></i>
-        <span>
-            {{ 'Log in as user'|trans }}
-        </span>
-    </a>
 {% endblock %}
