+++ packages/frontend-api/composer.json
@@ -23,6 +23,7 @@
     "require": {
         "php": "^8.3",
         "elasticsearch/elasticsearch": "^7.6.1",
+        "hybridauth/hybridauth": "^3.11",
         "lcobucci/jwt": "^4.1.5",
         "overblog/graphql-bundle": "1.3.2",
         "overblog/graphiql-bundle": "^0.3",
