+++ packages/frontend-api/src/DependencyInjection/ShopsysFrontendApiExtension.php
@@ -32,6 +32,13 @@ class ShopsysFrontendApiExtension extends Extension implements PrependExtensionI
         ];
 
         $container->prependExtensionConfig('overblog_graphql', $config);
+        $fileLocator = (new FileLocator(__DIR__ . '/../Resources/views/bundles'));
+
+        $container->loadFromExtension('twig', [
+            'paths' => [
+                $fileLocator->locate('ShopsysFrameworkBundle') => 'ShopsysFramework',
+            ],
+        ]);
     }
 
     /**
