+++ packages/frontend-api/src/Model/Customer/User/CustomerUserUpdateDataFactory.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace Shopsys\FrontendApiBundle\Model\Customer\User;
 
 use Overblog\GraphQLBundle\Definition\Argument;
+use Shopsys\FrameworkBundle\Model\Customer\BillingAddressDataFactory;
 use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser;
+use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserDataFactory as FrameworkCustomerUserDataFactory;
 use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateData;
 use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateDataFactoryInterface;
 
@@ -13,9 +15,13 @@ class CustomerUserUpdateDataFactory
 {
     /**
      * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateDataFactoryInterface $customerUserUpdateDataFactory
+     * @param \Shopsys\FrameworkBundle\Model\Customer\BillingAddressDataFactory $billingAddressDataFactory
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserDataFactory $customerUserDataFactory
      */
     public function __construct(
         protected readonly CustomerUserUpdateDataFactoryInterface $customerUserUpdateDataFactory,
+        protected readonly BillingAddressDataFactory $billingAddressDataFactory,
+        protected readonly FrameworkCustomerUserDataFactory $customerUserDataFactory,
     ) {
     }
 
@@ -38,6 +44,49 @@ class CustomerUserUpdateDataFactory
                 $customerUserData->{$key} = $value;
             }
         }
+
+        return $customerUserUpdateData;
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @return \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateData
+     */
+    public function createFromCustomerUser(CustomerUser $customerUser): CustomerUserUpdateData
+    {
+        return $this->customerUserUpdateDataFactory->createFromCustomerUser($customerUser);
+    }
+
+    /**
+     * @param \Shopsys\FrontendApiBundle\Model\Customer\User\RegistrationData $registrationData
+     * @return \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateData
+     */
+    public function createFromRegistrationData(RegistrationData $registrationData): CustomerUserUpdateData
+    {
+        $billingAddressData = $this->billingAddressDataFactory->create();
+        $billingAddressData->city = $registrationData->city;
+        $billingAddressData->street = $registrationData->street;
+        $billingAddressData->postcode = $registrationData->postcode;
+        $billingAddressData->country = $registrationData->country;
+        $billingAddressData->companyCustomer = $registrationData->companyCustomer;
+        $billingAddressData->companyName = $registrationData->companyName;
+        $billingAddressData->companyNumber = $registrationData->companyNumber;
+        $billingAddressData->companyTaxNumber = $registrationData->companyTaxNumber;
+        $billingAddressData->activated = $registrationData->activated;
+
+        $customerUserData = $this->customerUserDataFactory->createForDomainId($registrationData->domainId);
+        $customerUserData->createdAt = $registrationData->createdAt;
+        $customerUserData->email = $registrationData->email;
+        $customerUserData->lastName = $registrationData->lastName;
+        $customerUserData->password = $registrationData->password;
+        $customerUserData->firstName = $registrationData->firstName;
+        $customerUserData->telephone = $registrationData->telephone;
+        $customerUserData->newsletterSubscription = $registrationData->newsletterSubscription;
+
+        $customerUserUpdateData = $this->customerUserUpdateDataFactory->create();
+        $customerUserUpdateData->billingAddressData = $billingAddressData;
+        $customerUserUpdateData->customerUserData = $customerUserData;
+        $customerUserUpdateData->sendRegistrationMail = $registrationData->activated;
 
         return $customerUserUpdateData;
     }
