+++ packages/frontend-api/src/Model/Customer/User/RegistrationData.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shopsys\FrameworkBundle\Model\Customer\User;
+namespace Shopsys\FrontendApiBundle\Model\Customer\User;
 
 class RegistrationData
 {
