--- packages/frontend-api/src/Model/Customer/User/RegistrationDataApiFactory.php
@@ -1,48 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrontendApiBundle\Model\Customer\User;
-
-use Overblog\GraphQLBundle\Definition\Argument;
-use Shopsys\FrameworkBundle\Component\Domain\Domain;
-use Shopsys\FrameworkBundle\Model\Country\CountryFacade;
-use Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData;
-use Shopsys\FrameworkBundle\Model\Customer\User\RegistrationDataFactory;
-
-class RegistrationDataApiFactory
-{
-    /**
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
-     * @param \Shopsys\FrameworkBundle\Model\Country\CountryFacade $countryFacade
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\RegistrationDataFactory $registrationDataFactory
-     */
-    public function __construct(
-        protected readonly Domain $domain,
-        protected readonly CountryFacade $countryFacade,
-        protected readonly RegistrationDataFactory $registrationDataFactory,
-    ) {
-    }
-
-    /**
-     * @param \Overblog\GraphQLBundle\Definition\Argument $argument
-     * @return \Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData
-     */
-    public function createWithArgument(Argument $argument): RegistrationData
-    {
-        $input = $argument['input'];
-
-        $domainId = $this->domain->getId();
-        $registrationData = $this->registrationDataFactory->createForDomainId($domainId);
-
-        foreach ($input as $key => $value) {
-            if (property_exists(get_class($registrationData), $key)) {
-                $registrationData->{$key} = $value;
-            }
-        }
-
-        $registrationData->country = $this->countryFacade->findByCode($input['country']);
-
-        return $registrationData;
-    }
-}
