+++ packages/frontend-api/src/Model/Security/LoginAsUserFacade.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shopsys\FrameworkBundle\Model\Security;
+namespace Shopsys\FrontendApiBundle\Model\Security;
 
 use Ramsey\Uuid\Uuid;
 use Shopsys\FrameworkBundle\Model\Administrator\Administrator;
@@ -14,17 +14,11 @@ use Shopsys\FrameworkBundle\Model\Security\Exception\LoginAsRememberedUserExcept
 use Shopsys\FrontendApiBundle\Model\Token\TokenAuthenticator;
 use Shopsys\FrontendApiBundle\Model\Token\TokenFacade;
 use Shopsys\FrontendApiBundle\Model\User\FrontendApiUser;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class LoginAsUserFacade
 {
-    public const SESSION_LOGIN_AS = 'loginAsUser';
-
     /**
-     * @param \Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface $tokenStorage
-     * @param \Symfony\Component\EventDispatcher\EventDispatcherInterface $eventDispatcher
      * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserRepository $customerUserRepository
      * @param \Shopsys\FrameworkBundle\Model\Administrator\Security\AdministratorFrontSecurityFacade $administratorFrontSecurityFacade
      * @param \Symfony\Component\HttpFoundation\RequestStack $requestStack
@@ -33,8 +27,6 @@ class LoginAsUserFacade
      * @param \Shopsys\FrontendApiBundle\Model\Token\TokenFacade $tokenFacade
      */
     public function __construct(
-        protected readonly TokenStorageInterface $tokenStorage,
-        protected readonly EventDispatcherInterface $eventDispatcher,
         protected readonly CustomerUserRepository $customerUserRepository,
         protected readonly AdministratorFrontSecurityFacade $administratorFrontSecurityFacade,
         protected readonly RequestStack $requestStack,
