+++ packages/frontend-api/src/Resources/config/services.yaml
@@ -30,8 +30,14 @@ services:
 
     Shopsys\FrontendApiBundle\Component\GqlContext\GqlContextInitializer: ~
 
+    Shopsys\FrontendApiBundle\Controller\CustomerUserController:
+        tags: [ 'controller.service_arguments' ]
+
     Shopsys\FrontendApiBundle\Controller\FrontendApiController:
         tags: ['controller.service_arguments']
+
+    Shopsys\FrontendApiBundle\Controller\SocialNetworkController:
+        tags: [ 'controller.service_arguments' ]
 
     Shopsys\FrontendApiBundle\Model\Error\ErrorHandlerListener:
         tags:
@@ -82,6 +88,14 @@ services:
         arguments:
             $productSearchResultsProviders: !tagged 'shopsys.frontend_api.products_search_results_provider'
 
+    Shopsys\FrontendApiBundle\Model\SocialNetwork\SocialNetworkConfigFactory:
+        arguments:
+            $socialNetworkLoginConfig: '%social_network_login_config%'
+
+    Shopsys\FrontendApiBundle\Model\SocialNetwork\SocialNetworkFacade:
+        arguments:
+            $logger: '@monolog.logger.socialNetwork'
+
     Shopsys\FrontendApiBundle\Model\Token\TokenAuthenticator:  ~
 
     Shopsys\FrontendApiBundle\Model\User\FrontendApiUserFactory: ~
@@ -90,3 +104,5 @@ services:
         alias: Shopsys\FrontendApiBundle\Model\User\FrontendApiUserFactory
 
     Shopsys\FrontendApiBundle\Model\User\FrontendApiUserProvider: ~
+
+    Shopsys\FrontendApiBundle\Twig\LoginAsUserExtension: ~
