+++ packages/frontend-api/src/Resources/views/bundles/ShopsysFrameworkBundle/Admin/Content/Customer/edit.html.twig
@@ -1,0 +1,10 @@
+{% extends '@!ShopsysFramework/Admin/Content/Customer/edit.html.twig' %}
+{%  block h1 %}
+    {{ parent() }}
+    <a href="{{ getSsoLoginAsCustomerUserUrl(customerUser) }}" target="_blank" class="btn-link-style btn-link-style--iconic-l">
+        <i class="svg svg-forward-page"></i>
+        <span>
+            {{ 'Log in as user'|trans }}
+        </span>
+    </a>
+{% endblock %}
