+++ project-base/app/config/services.yaml
@@ -111,10 +111,6 @@ services:
 
     App\Model\Customer\User\CurrentCustomerUser: ~
 
-    App\Model\Customer\User\RegistrationDataFactory: ~
-
-    App\Model\Customer\User\RegistrationFacade: ~
-
     App\Model\LanguageConstant\LanguageConstantFacade:
         arguments:
             $languageConstantsUrlPattern: '%env(LANGUAGE_CONSTANTS_URL_PATTERN)%'
@@ -359,12 +355,6 @@ services:
     Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateDataFactoryInterface:
         alias: App\Model\Customer\User\CustomerUserUpdateDataFactory
 
-    Shopsys\FrameworkBundle\Model\Customer\User\RegistrationDataFactory:
-        alias: App\Model\Customer\User\RegistrationDataFactory
-
-    Shopsys\FrameworkBundle\Model\Customer\User\RegistrationFacade:
-        alias: App\Model\Customer\User\RegistrationFacade
-
     Shopsys\FrameworkBundle\Model\Feed\FeedFacade:
         alias: App\Model\Feed\FeedFacade
 
@@ -571,9 +561,6 @@ services:
 
     Shopsys\FrameworkBundle\Model\Product\Search\ProductFilterDataToQueryTransformer:
         alias: App\Model\Product\Search\ProductFilterDataToQueryTransformer
-
-    Shopsys\FrameworkBundle\Model\Security\LoginAsUserFacade:
-        alias: App\Model\Security\LoginAsUserFacade
 
     Shopsys\FrameworkBundle\Model\Sitemap\SitemapFacade:
         alias: App\Model\Sitemap\SitemapFacade
