+++ project-base/app/src/Model/Customer/User/CustomerUserUpdateDataFactory.php
@@ -21,7 +21,6 @@ use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserUpdateDataFactory as
  * @method \App\Model\Customer\User\CustomerUserUpdateData createFromCustomerUser(\App\Model\Customer\User\CustomerUser $customerUser)
  * @method \App\Model\Customer\User\CustomerUserUpdateData createAmendedByOrder(\App\Model\Customer\User\CustomerUser $customerUser, \App\Model\Order\Order $order, \App\Model\Customer\DeliveryAddress|null $deliveryAddress)
  * @method \App\Model\Customer\BillingAddressData getAmendedBillingAddressDataByOrder(\App\Model\Order\Order $order, \App\Model\Customer\BillingAddress $billingAddress)
- * @method \App\Model\Customer\User\CustomerUserUpdateData createFromRegistrationData(\App\Model\Customer\User\RegistrationData $registrationData)
  */
 class CustomerUserUpdateDataFactory extends BaseCustomerUserUpdateDataFactory
 {
