--- project-base/app/src/Model/Customer/User/RegistrationData.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Model\Customer\User;
-
-use Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData as BaseRegistrationData;
-
-class RegistrationData extends BaseRegistrationData
-{
-}
