+++ upgrade-notes/backend_20240529_084428.md
@@ -994,6 +994,7 @@
 
 -   implemented are login via Facebook, Google and Seznam
 -   add url `{eshop_domain}/social-network/login/{type}` as redirect url in you social network. `type` is implemented service e.g.: `google`, `facebook`, `seznam`
+-   `Shopsys\FrameworkBundle\Model\Administrator\Security\AdministratorFrontSecurityFacade::isAdministratorLoggedAsCustomer()` method was removed
 -   `Shopsys\FrameworkBundle\Controller\Admin\AdministratorController` constructor was changed:
     ```diff
         public function __construct(
@@ -1001,8 +1002,18 @@
     +       protected readonly AdministratorTwoFactorAuthenticationFacade $administratorTwoFactorAuthenticationFacade,
     ```
 -   `Shopsys\FrameworkBundle\Controller\Admin\CustomerController` class was changed:
+    -   `__construct()` method changed its interface:
+    ```diff
+        public function __construct(
+            // ...
+            protected readonly OrderFacade $orderFacade,
+    -       protected readonly LoginAsUserFacade $loginAsUserFacade,
+            protected readonly DomainRouterFactory $domainRouterFactory,
+            // ...
+    ```
     -   `LOGIN_AS_TOKEN_ID_PREFIX` constant was removed
-    -   `loginAsUserAction()` method was removed, use `loginAsCustomerUserAction` instead
+    -   `loginAsUserAction()` method was removed, use `Shopsys\FrontendApiBundle\Controller\CustomerUserController::loginAsCustomerUserAction()` instead
+    -   `getSsoLoginAsCustomerUserUrl()` method was removed, use `Shopsys\FrontendApiBundle\Twig\LoginAsUserExtension::getSsoLoginAsCustomerUserUrl()` instead
 -   `Shopsys\FrameworkBundle\Model\Administrator\Administrator` class was changed:
     -   the class now implements `Scheb\TwoFactorBundle\Model\Email\TwoFactorInterface` and `Scheb\TwoFactorBundle\Model\Google\TwoFactorInterface`
     -   `checkRolesContainAdminRole()` method was removed
@@ -1013,20 +1024,20 @@
     +       protected readonly CustomerActivationMail $customerActivationMail,
     ```
 -   `Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserFacade` class was changed:
--   `__construct()` changed its interface:
-    ```diff
-        public function __construct(
-            // ...
-    +       protected readonly NewsletterFacade $newsletterFacade,
-    +       protected readonly HashGenerator $hashGenerator,
-    ```
--   `edit()` function is now public
--   `addRefreshTokenChain()` method changed its interface:
-    ```diff
-        public function addRefreshTokenChain(
-            // ...
-    +       ?Administrator $administrator,
-    ```
+    -   `__construct()` changed its interface:
+        ```diff
+            public function __construct(
+                // ...
+        +       protected readonly NewsletterFacade $newsletterFacade,
+        +       protected readonly HashGenerator $hashGenerator,
+        ```
+    -   `edit()` function is now public
+    -   `addRefreshTokenChain()` method changed its interface:
+        ```diff
+            public function addRefreshTokenChain(
+                // ...
+        +       ?Administrator $administrator,
+        ```
 -   `Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserRefreshTokenChainFacade` class was changed:
     -   `createCustomerUserRefreshTokenChain()` method changed its interface:
         ```diff
@@ -1049,7 +1060,7 @@
                 // ...
         +       protected readonly Localization $localization,
         ```
--   `Shopsys\FrontendApiBundle\Model\Mutation\Customer\User\CustomerUserMutation` class was chagned:
+-   `Shopsys\FrontendApiBundle\Model\Mutation\Customer\User\CustomerUserMutation` class was changed:
     -   `__contruct()` method changed its interface:
         ```diff
             public function __contruct(
@@ -1058,7 +1069,6 @@
         +       protected readonly RegistrationDataFactory $registrationDataFactory,
         +       protected readonly MergeCartFacade $mergeCartFacade,
         +       protected readonly OrderApiFacade $orderFacade,
-        +       protected readonly RegistrationDataApiFactory $registrationDataApiFactory,
         ```
     -   `registerMutation()` - the returned array changed its structure to `{tokens: array{accessToken: string, refreshToken: string}, showCartMergeInfo: bool}`
 -   `Shopsys\FrontendApiBundle\Model\Order\OrderApiFacade::__contruct()` changed its interface:
@@ -1087,8 +1097,17 @@
             // ...
     +       protected readonly ?string $administratorUuid,
     ```
+-   `Shopsys\FrameworkBundle\Model\Security\LoginAsUserFacade` was removed, use `Shopsys\FrontendApiBundle\Model\Security\LoginAsUserFacade` instead
+-   `Shopsys\FrontendApiBundle\Model\Customer\User\CustomerUserUpdateDataFactory::__construct()` changed its interface:
+
+    ```diff
+         public function __construct(
+            protected readonly CustomerUserUpdateDataFactoryInterface $customerUserUpdateDataFactory,
+    +       protected readonly BillingAddressDataFactory $billingAddressDataFactory,
+    +       protected readonly \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserDataFactory $customerUserDataFactory,
+    ```
+
 -   [features moved](#movement-of-features-from-project-base-to-packages) from project-base to the framework package:
-    -   `CustomerController::loginAsCustomerUserAction()` and all the related logic (see `LoginAsUserFacade`)
     -   `AdministratorRoleGroup` entity, `Administrator::$roleGroup` and all the related logic
     -   `Roles`, `RolesType`, `MenuItemsGrantedRolesSetting`, and all the related logic
         -   removing not granted items from the menu is now moved from `SideMenuConfigurationSubscriber` to the new `MenuItemsGrantedRolesSubscriber` class
@@ -1100,17 +1119,16 @@
     -   `CustomerUser::$newsletterSubscription` and all the related logic
     -   `CustomerUserRefreshTokenChain::$administrator`
     -   `RemoveOldCustomerUserRefreshTokenChainsCronModule` and all the related logic
-    -   `CustomerUserUpdateDataFactory` methods:
-        -   `createAmendedByOrder()`
-        -   `createFromRegistrationData()`
-    -   `RegistrationData` class
-    -   `RegistrationDataFactory` class
-    -   `RegistrationFacade` class
+    -   `CustomerUserUpdateDataFactory::createAmendedByOrder()`
     -   `MailTemplate::$orderStatus` property and all the related logic
     -   `MailTemplateBuilder` class
     -   `MailSetting` class constants `MAIL_FACEBOOK_URL`, `MAIL_INSTAGRAM_URL`, `MAIL_YOUTUBE_URL`, `MAIL_LINKEDIN_URL`, `MAIL_TIKTOK_URL`, and `MAIL_FOOTER_TEXT` and all the related logic
     -   `Order::setCustomerUser()` method
 -   [features moved](#movement-of-features-from-project-base-to-packages) from project-base to the frontend-api package:
+    -   `CustomerController::loginAsCustomerUserAction()` and all the related logic
+        -   `Shopsys\FrontendApiBundle\Model\Security\LoginAsUserFacade`
+        -   `admin_customer_loginascustomeruser` route was renamed to `admin_customeruser_loginascustomeruser`
+    -   `CustomerUserUpdateDataFactory::createFromRegistrationData()`
     -   `OrderApiFacade::ONE_HOUR_REGISTRATION_WINDOW` constant
     -   `OrderApiFacade` methods:
         -   `getByOrderNumberAndCustomerUser()`
@@ -1120,6 +1138,9 @@
         -   `getByOrderNumberAndCustomerUser()`
         -   `findByOrderNumberAndCustomerUser()`
         -   `getByUuidAndCustomerUser()`
+    -   `RegistrationData` class
+    -   `RegistrationDataFactory` class
+    -   `RegistrationFacade` class
     -   `TokenAuthenticator::HEADER_AUTHORIZATION` constant
     -   `TokenAuthenticator` methods:
         -   `getCredentials()`
